Added placeholder for URL, Removed error message in registration on focus, Cleaned code, Added package and schema for ERD, Added country code type for DEFAULT_COUNTRY in readme, Added Username: prefix for the registration page errors, Checking cookie for all end points, Removed admin end points

+++ src/routes/(product)/deleteProduct/[name]/+page.server.ts
@@ -1,53 +1,50 @@
 import type { PageServerLoad } from "./$types";
 import { redirect } from "@sveltejs/kit";
 import { db } from "$lib/database";
-import type { Action, Actions } from "./$types";
-import type Page from "../../../+page.svelte";
 
-//
 export const load = (async ({ params: { name }, locals }) => {
-  //console.log("delete name: ", name);
-  const deleteProduct = await db.product.findUnique({
-    where: {
-      name_owner: {
-        name: name,
-        owner: locals.user.name,
-      },
-    },
-    select: {
-      id: true,
-      name: true,
-
-      links: {
-        select: {
-          country: true,
-          url: true,
+  if (locals.user) {
+    const deleteProduct = await db.product.findUnique({
+      where: {
+        name_owner: {
+          name: name,
+          owner: locals.user.name,
         },
       },
-    },
-  });
-  //console.log("delete product:::", deleteProduct);
-  const removedLinks = await db.product.update({
-    where: {
-      id: deleteProduct?.id,
-    },
-    data: {
-      links: {
-        deleteMany: {},
+      select: {
+        id: true,
+        name: true,
+
+        links: {
+          select: {
+            country: true,
+            url: true,
+          },
+        },
       },
-    },
-    include: {
-      links: true,
-    },
-  });
-  //console.log("removedLinks ", removedLinks);
+    });
+    const removedLinks = await db.product.update({
+      where: {
+        id: deleteProduct?.id,
+      },
+      data: {
+        links: {
+          deleteMany: {},
+        },
+      },
+      include: {
+        links: true,
+      },
+    });
 
-  const deletedProduct = await db.product.delete({
-    where: {
-      id: deleteProduct?.id,
-    },
-  });
-  //console.log("Deleted Product ", deletedProduct);
+    const deletedProduct = await db.product.delete({
+      where: {
+        id: deleteProduct?.id,
+      },
+    });
 
-  throw redirect(302, "/productsList");
+    throw redirect(302, "/productsList");
+  } else {
+    throw redirect(303, "/");
+  }
 }) satisfies PageServerLoad;
